move validations into controller
